Passing values to the ports checker
